Extract Planning runtime packets
